requirements.txt
# Core dependencies
click>=8.2.1
faiss-cpu>=1.12.0
numpy>=2.3.2
PyMuPDF>=1.26.3
psutil>=7.0.0
requests>=2.32.4
sentence-transformers>=5.1.0
tqdm>=4.67.1

# Dependencies for sentence-transformers
torch>=2.8.0
transformers>=4.55.2
